moved `showdown` config and extensions to a new folder
(`shared` as folder name is to be decided/discussed, for now it’s just a way to move it somewhere else)

(`shared` as folder name is to be decided/discussed, for now it’s just a way to move it somewhere else)
